move vm code to widget
