+++ lib/pkg/init_dispose.dart
@@ -1,0 +1,1 @@
+typedef InitDisposeFn = (void Function(), void Function());
