--- lib/screen/resources_create_capture_vm.dart
@@ -1,52 +1,0 @@
-import 'package:flutter/foundation.dart';
-import 'package:open_scanner/domain/image.dart';
-import 'package:open_scanner/pkg/safe_catch.dart';
-import 'package:open_scanner/repo/crop_tool.dart';
-import 'package:camera/camera.dart';
-
-class ResourcesCreateCaptureVM {
-  final CropToolRepo cropToolRepo;
-
-  ResourcesCreateCaptureVM(this.cropToolRepo);
-
-  Future<CameraController> getCameraController() async {
-    final cameras = await availableCameras();
-    final camera = cameras.firstWhere(
-      (camera) => camera.lensDirection == CameraLensDirection.back,
-      orElse: () => cameras.first,
-    );
-    final controller = CameraController(
-      camera,
-      ResolutionPreset.ultraHigh,
-      enableAudio: false,
-    );
-
-    await controller.initialize();
-    return controller;
-  }
-
-  disposeController(CameraController controller) {
-    controller.dispose();
-  }
-
-  Future<XFile> takePicture(CameraController controller) async {
-    await safeCatchFuture(() => controller.setFlashMode(FlashMode.off));
-    await safeCatchFuture(() => controller.setFocusMode(FocusMode.locked));
-    await safeCatchFuture(
-        () => controller.setExposureMode(ExposureMode.locked));
-    final imageFile = await controller.takePicture();
-    await safeCatchFuture(() => controller.setFocusMode(FocusMode.auto));
-    await safeCatchFuture(() => controller.setExposureMode(ExposureMode.auto));
-
-    return imageFile;
-  }
-
-  Future<Error?> setSelectedImage(XFile file) async {
-    final bytes = await file.readAsBytes();
-    final img = await ImageDomain.fromEncodedList(bytes);
-    if (img == null) return FlutterError("Decoded image returns null");
-
-    cropToolRepo.selectedImage = img;
-    return null;
-  }
-}
