+++ lib/screen/resources_create_select.dart
@@ -1,12 +1,49 @@
 import 'package:open_scanner/component/crop_tool.dart';
-import 'package:open_scanner/screen/resources_create_select_vm.dart';
+import 'package:open_scanner/domain/image.dart';
+import 'package:open_scanner/pkg/init_dispose.dart';
+import 'package:open_scanner/repo/crop_tool.dart';
+import 'package:open_scanner/repo/resource.dart';
 import 'package:flutter/material.dart';
 
 class ResourcesCreateSelectScreen extends StatefulWidget {
-  final ResourcesCreateSelectVM vm;
+  final CropToolRepo cropToolRepo;
+  final ResourceRepo resourceRepo;
 
-  ResourcesCreateSelectScreen({super.key, required this.vm}){
-    vm.resetCropTool();
+  void resetCropTool() {
+    cropToolRepo.resetCropTools();
+    resourceRepo.resetResources();
+    
+    cropToolRepo.addCropTool(const Offset(300, 300));
+    resourceRepo.addResource();
+  }
+
+  ImageDomain getSelectedImage() {
+    return cropToolRepo.selectedImage;
+  }
+
+  int getCropToolLength() {
+    return cropToolRepo.entities.length;
+  }
+
+  InitDisposeFn getCropToolListener(){
+    final List<InitDisposeFn> initDisposeFns = [];
+    
+    for (var i = 0; i < cropToolRepo.entityNotifiers.length; i++) {
+      fn() => resourceRepo.resetResource(i);
+      initDisposeFns.add((
+          () { cropToolRepo.entityNotifiers[i].addListener(fn); },
+          () { cropToolRepo.entityNotifiers[i].removeListener(fn); },
+      ));
+    }
+
+    return (
+      () { for (var (initFn, _) in initDisposeFns) { initFn(); } },
+      () { for (var (_, disposeFn) in initDisposeFns) { disposeFn(); } },
+    );
+  }
+
+  ResourcesCreateSelectScreen(this.cropToolRepo, this.resourceRepo, {super.key}){
+    resetCropTool();
   }
 
   @override
@@ -17,8 +54,8 @@ class ResourcesCreateSelectScreen extends StatefulWidget {
 // A widget that displays the picture taken by the user.
 class ResourcesCreateSelectScreenState
     extends State<ResourcesCreateSelectScreen> {
-  late InitDisposeFn cropToolListener = widget.vm.getCropToolListener();
-  late final image = widget.vm.getSelectedImage(); 
+  late InitDisposeFn cropToolListener = widget.getCropToolListener();
+  late final image = widget.getSelectedImage(); 
 
   @override
   void initState() {
@@ -69,8 +106,8 @@ class ResourcesCreateSelectScreenState
                   );
                 }),
             // Container(width: 100, height: 100, color: Colors.red),
-            for (var i = 0; i < widget.vm.getCropToolLength(); i++)
-              SizedBox(width: image.size.width, height: image.size.height, child: CropTool(i, widget.vm.cropToolRepo))
+            for (var i = 0; i < widget.getCropToolLength(); i++)
+              SizedBox(width: image.size.width, height: image.size.height, child: CropTool(i, widget.cropToolRepo))
           ],
         ),
       ),
